Cache locations for neighbours and incident edges.
This works by offering a dedicated cache location for
each individual set of tail/head neighbours.  The cost
is a greater memory footprint for an individual graph,
the benefit is that the cache is "safe" and reliable
(i.e. it won't overwrite another vertex's neighbours)
provided the graph itself is static (i.e. you don't
add more vertices or edges).

It could in principle be extended with a cache check
which notes if the cache needs to be updated, to save
the recalculation of the neighbours.

The lack of safety in the event of adding/removing
vertices and edges is an annoyance, and some thought
needs to be given to how to address this.