VertexRange type for range-based access to vertex properties
This new type is designed for the construction of RandomAccessRanges
of graph vertex properties, where each individual element of the range
is the given property for the corresponding individual vertex.

The current patch makes use of this new type to implement new versions
of the neighbours() and incidentEdges() graph methods.