Correct use of 'tail' and 'head' vertex terminology
The previous use of 'tail' and 'head' in dgraph was in opposition to the
standard usage in graph theory: 'tail' means the vertex an edge leads
from, whereas 'head' means the vertex an edge leads to.

This patch consequently inverts the use of these terms.  This should not
affect downstream use of the code; public methods are unaltered, except
in terms of their parameter names, which should not impact usage.

As a by-product, the naming conventions for the reworked private
variables have been changed _fromThis toThis_ to reflect concerns that,
at least in principle, names with leading underscores may be reserved.

Reference for the head/tail change includes Wikipedia:
https://en.wikipedia.org/wiki/Graph_arc#Basic_terminology

... and the igraph documentation of its edges' 'from' and 'to'
parameters:
http://igraph.org/c/doc/igraph-Basic.html#igraph_edge

Fixes WebDrake/Dgraph#8.