Add module of graph property utilities
Graph properties are custom data fields that may be associated with any
graph elements of choice: edges, vertices, or the graph itself as a
whole.  They provide the basis for much extensibility of the basic edge,
vertex and graph data structures.

This module offers utilities for generating graph properties themselves,
together with mixin templates that can be used to generate the variables
in which to store the values of individual properties.  Such variables
will also be tagged with graph-property user-defined attributes (UDAs)
in order to clearly mark them out as containing graph property values.